+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * [#997](https://github.com/xmake-io/xmake/issues/997): Support to set std lanuages for `xmake project -k cmake`
 * [#998](https://github.com/xmake-io/xmake/issues/998): Support to install vcpkg packages with windows-static-md
+* [#996](https://github.com/xmake-io/xmake/issues/996): Improve to find vcpkg directory
 
 ## v2.3.8
 
@@ -851,6 +852,7 @@
 
 * [#997](https://github.com/xmake-io/xmake/issues/997): `xmake project -k cmake` 插件增加对 `set_languages` 的支持
 * [#998](https://github.com/xmake-io/xmake/issues/998): 支持安装 windows-static-md 类型的 vcpkg 包
+* [#996](https://github.com/xmake-io/xmake/issues/996): 改进 vcpkg 目录查找
 
 ## v2.3.8